requirements.txt
# Django and REST Framework
Django==4.2.7
djangorestframework==3.14.0

# Database
psycopg2-binary==2.9.9

# Filtering and pagination
django-filter==23.5

# CORS headers
django-cors-headers==4.3.1

# Environment variables
python-decouple==3.8

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0

# Code quality
flake8==6.1.0
black==23.12.1
isort==5.13.2

# Documentation
drf-spectacular==0.27.0

# Development tools
ipython==8.18.1
django-extensions==3.2.3

